Moved draw call class hierarchy into Renderer class. Added a FEATURES.md to track version features in engine module. Updated javadocs of Animation and AnimationSlice classes
