--- engine/src/main/java/dev/gamekit/graphics/DrawLine.java
@@ -1,20 +1,0 @@
-package dev.gamekit.graphics;
-
-import java.awt.*;
-
-/** {@link DrawLine} is a draw call which renders a line */
-public class DrawLine extends DrawShape<DrawLine> {
-  private final int x1, y1, x2, y2;
-
-  public DrawLine(int x1, int y1, int x2, int y2) {
-    this.x1 = x1;
-    this.y1 = y1;
-    this.x2 = x2;
-    this.y2 = y2;
-  }
-
-  @Override
-  protected void draw(Graphics2D g) {
-    g.drawLine(x1, -y1, x2, -y2);
-  }
-}
