--- engine/src/main/java/dev/gamekit/graphics/DrawRect.java
@@ -1,8 +1,0 @@
-package dev.gamekit.graphics;
-
-/** {@link DrawRect} is a draw call which renders a <b>center-origin</b> rect */
-public class DrawRect extends DrawRoundRect {
-  public DrawRect(int x, int y, int width, int height, boolean filled) {
-    super(x, y, width, height, 0, 0, filled);
-  }
-}
